Moved classes to package to organize them
